feature: 布局模板预览
