+++ paas-ce/lesscode/lib/client/src/App.vue
@@ -60,7 +60,7 @@
         computed: {
             ...mapGetters(['mainContentLoading']),
             emptyPage () {
-                return this.$route.name === 'preview' || this.$route.name === 'previewTemplate' || this.$route.name === 'previewLayout'
+                return this.$route.name === 'preview' || this.$route.name === 'previewTemplate'
             },
             authed () {
                 return this.$route.meta.authed
