+++ paas-ce/lesscode/lib/client/src/components/project/layout-thumb-list.vue
@@ -59,7 +59,7 @@
                 return require(`@/images/${previewImg}`)
             },
             handlePreview (layout) {
-                window.open(`/preview-layout/project/${layout.projectId}/${layout.id}`, '_blank')
+                window.open(`/preview-template/project/${layout.projectId}/${layout.id}?type=nav-template`, '_blank')
             }
         }
     }
